@@ -1,6 +1,6 @@
 {
   "name": "rangeslider-js",
-  "version": "3.1.2",
+  "version": "3.2.0",
   "main": "./dist/rangeslider-js.min.js",
   "description": "lightweight range slider",
   "author": "Steffen Bär <stbaer.mail@gmail.com>",
@@ -39,37 +39,38 @@
     "ev-pos": "^1.0.1"
   },
   "devDependencies": {
-    "autoprefixer": "8.3.0",
+    "@babel/core": "7.1.6",
+    "@babel/preset-env": "7.1.6",
+    "@babel/preset-stage-2": "7.0.0",
+    "autoprefixer": "9.3.1",
     "ava": "0.25.0",
-    "babel-core": "6.26.0",
-    "babel-eslint": "8.2.3",
-    "babel-loader": "7.1.4",
-    "babel-preset-env": "1.6.1",
+    "babel-eslint": "10.0.1",
+    "babel-loader": "8.0.4",
     "browser-env": "3.2.5",
-    "css-loader": "0.28.11",
-    "cssnano": "^3.10.0",
-    "eslint": "4.19.1",
-    "eslint-config-standard": "11.0.0",
+    "css-loader": "1.0.1",
+    "csso-cli": "1.1.0",
+    "eslint": "5.9.0",
+    "eslint-config-standard": "12.0.0",
     "eslint-friendly-formatter": "4.0.1",
-    "eslint-loader": "2.0.0",
-    "eslint-plugin-import": "2.11.0",
-    "eslint-plugin-node": "6.0.1",
-    "eslint-plugin-promise": "3.7.0",
-    "eslint-plugin-standard": "3.1.0",
-    "file-loader": "1.1.11",
+    "eslint-loader": "2.1.1",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-node": "8.0.0",
+    "eslint-plugin-promise": "4.0.1",
+    "eslint-plugin-standard": "4.0.0",
+    "file-loader": "2.0.0",
     "friendly-errors-webpack-plugin": "1.7.0",
     "gf-release": "1.4.5",
     "html-webpack-plugin": "3.2.0",
     "mkdirp": "^0.5.1",
     "nodes-each": "1.0.3",
     "progress-bar-webpack-plugin": "1.11.0",
-    "style-loader": "0.21.0",
-    "url-loader": "1.0.1",
-    "webpack": "4.6.0",
-    "webpack-bundle-analyzer": "2.11.1",
-    "webpack-cli": "2.0.15",
-    "webpack-dev-server": "3.1.3",
-    "webpack-merge": "4.1.2"
+    "style-loader": "0.23.1",
+    "url-loader": "1.1.2",
+    "webpack": "4.26.1",
+    "webpack-bundle-analyzer": "3.0.3",
+    "webpack-cli": "3.1.2",
+    "webpack-dev-server": "3.1.10",
+    "webpack-merge": "4.1.4"
   },
   "files": [
     "dist/",
